Some things are working
Working insert and handles splitting the node if it overflows. The
second insert after this hangs the program. The tree can be output
validating the split. There is now functionality to calculate the MBR
from a list of points (entries). Max and min within this function are
currently doubles, this locks the tree class to numerical data, this
might not be a problem.